requirements.txt
confluent-kafka==2.6.1
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
scikit-learn==1.5.1
lightgbm==4.5.0
mlflow==2.12.1
prometheus-client==0.20.0
joblib==1.4.2
pyarrow==15.0.2